Refactor page validation
