+++ src/app/[countryCode]/community/[[...page]]/page.tsx
@@ -1,7 +1,6 @@
 import { flatten, times } from 'lodash-es'
 import { Metadata } from 'next'
 import { notFound } from 'next/navigation'
-import { z } from 'zod'
 
 import { RecentActivityAndLeaderboardTabs } from '@/components/app/pageHome/recentActivityAndLeaderboardTabs'
 import {
@@ -11,6 +10,7 @@ import {
   PageLeaderboardInferredProps,
 } from '@/components/app/pageLeaderboard'
 import { COMMUNITY_PAGINATION_DATA } from '@/components/app/pageLeaderboard/constants'
+import { validatePageNum } from '@/components/app/pageLeaderboard/pageValidator'
 import { getPublicRecentActivity } from '@/data/recentActivity/getPublicRecentActivity'
 import { PageProps } from '@/types'
 import { generateMetadataDetails } from '@/utils/server/metadataUtils'
@@ -26,18 +26,6 @@ export async function generateMetadata(_props: Props): Promise<Metadata> {
     title: PAGE_LEADERBOARD_TITLE,
     description: PAGE_LEADERBOARD_DESCRIPTION,
   })
-}
-
-const pageValidator = z.string().pipe(z.coerce.number().int().gte(1).lte(50))
-const validatePageNum = ([page]: (string | undefined)[]) => {
-  if (!page) {
-    return 1
-  }
-  const val = pageValidator.safeParse(page)
-  if (val.success) {
-    return val.data
-  }
-  return null
 }
 
 // pre-generate the first 10 pages. If people want to go further, we'll generate them on the fly
